+++ CHANGELOG.md
@@ -11,9 +11,13 @@ this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - `RespondWith(this TestableHttpMessageHandler, HttpResponseMessage)` has been removed, the response is modified with every call, so it doesn't work reliably and is different from how HttpClientHandler works, which creates a HttpResponseMessage for every request.
 - `HttpResponseMessageBuilder` and `RespondWith(this TestableHttpMessageHandler, HttpResponseMessageBuilder)` has been removed, it's functionality can be replaced with ConfiguredResponse or a custom IResponse.
 
-## Changed
+### Added
+- Uri patterns now support query parameters.
+
+### Changed
 - Use the same parser for the assertion methods `WithRequestUri` (which is used by `ShouldHaveMadeRequestsTo`) as for the RoutingResponse functionality.
 - `RouteParserException` has been renamed to `UriPatternParserException`.
+- Renamed `RoutingOptions` to `UriPatternMatchingOptions`
 
 ## [0.9] - 2022-11-25
 ### Deprecated